Add files via upload
We postulate a panel data model for our study. Panel data models allow us to control for individual airline behaviour that is not quantifiable (that may be airport-invariant or timeinvariant), but may affect airline revenue (like cultural factors or difference in business practices across airlines). This individual airline heterogeneity, cannot be controlled for by a simple cross sectional or times series analysis.